Rename 'aux' dirs to 'auxil'.
