+++ clients/gtest/auxil/auxiliary_gtest.cpp
