+++ clients/gtest/auxil/set_get_mode_gtest.yaml
