+++ clients/include/auxil/testing_set_get_atomics_mode.hpp
