+++ clients/include/auxil/testing_set_get_pointer_mode.hpp
